removing duplicate code.
