+++ src/components/graphline.js
@@ -11,6 +11,7 @@ import {
 } from 'victory'
 import { useAppContext } from '../services/context'
 import { toTitleCase } from '../utils/strings'
+import { graphStyles } from '../utils/styles'
 
 /**
  * This function take a complete user object and complete
@@ -109,11 +110,13 @@ const GraphLine = ({
             // }}
             theme={VictoryTheme.material}
         >
-            {typeof title !== 'undefined' ? <VictoryLabel x={50} y={30} text={title} style={styles.title} /> : null}
+            {typeof title !== 'undefined' ? (
+                <VictoryLabel x={50} y={30} text={title} style={graphStyles.title} />
+            ) : null}
             <VictoryAxis
                 // fixLabelOverlap={true}
                 // scale={{ x: 'time' }}
-                style={styles.axisZero}
+                style={graphStyles.axisZero}
                 tickValues={xTickValues}
                 tickLabelComponent={<VictoryLabel text={({ datum }) => plot[datum]?.dow || datum} />}
             />
@@ -122,7 +125,7 @@ const GraphLine = ({
                 orientation="left"
                 tickValues={[0.2, 0.4, 0.6, 0.8, 1]}
                 tickFormat={t => (t * maxY1).toFixed(1)}
-                style={styles.axisOne}
+                style={graphStyles.axisOne}
             />
             {typeof line2 !== 'undefined' ? (
                 <VictoryAxis
@@ -131,7 +134,7 @@ const GraphLine = ({
                     maxDomain={maxY2}
                     tickValues={[0.2, 0.4, 0.6, 0.8, 1]}
                     tickFormat={t => (t * maxY2).toFixed(1)}
-                    style={styles.axisTwo}
+                    style={graphStyles.axisTwo}
                 />
             ) : null}
             {typeof line2 !== 'undefined' ? (
@@ -148,10 +151,10 @@ const GraphLine = ({
                         data-testid="line2"
                         interpolation={'natural'}
                         // labels={({ datum }) => (line2labels ? datum.y2.toFixed(0) : '')}
-                        style={styles.lineTwo}
+                        style={graphStyles.lineTwo}
                         labelComponent={<VictoryLabel dy={12} />}
                     />
-                    <VictoryScatter style={styles.seriesTwo} size={2} symbol="square" />
+                    <VictoryScatter style={graphStyles.seriesTwo} size={2} symbol="square" />
                 </VictoryGroup>
             ) : null}
             <VictoryGroup
@@ -166,9 +169,9 @@ const GraphLine = ({
                     data-testid="line1"
                     interpolation={'natural'}
                     // labels={({ datum }) => (line1labels ? datum.y1.toFixed(0) : '')}
-                    style={styles.lineOne}
+                    style={graphStyles.lineOne}
                 />
-                <VictoryScatter style={styles.seriesOne} size={2} />
+                <VictoryScatter style={graphStyles.seriesOne} size={2} />
             </VictoryGroup>
             <VictoryLegend
                 x={40}
@@ -177,15 +180,23 @@ const GraphLine = ({
                 orientation="horizontal"
                 gutter={30}
                 style={{
-                    border: { stroke: colorScale[0], strokeWidth: 0 },
-                    title: { ...baseLabelStyles, fontSize: 14 },
-                    labels: { ...baseLabelStyles, fontSize: 14 },
+                    border: { stroke: graphStyles.colorScale[0], strokeWidth: 0 },
+                    title: { ...graphStyles.baseLabelStyles, fontSize: 14 },
+                    labels: { ...graphStyles.baseLabelStyles, fontSize: 14 },
                     data: { fillOpacity: 1 },
                 }}
-                colorScale={colorScale}
+                colorScale={graphStyles.colorScale}
                 data={[
-                    { name: toTitleCase(line1), symbol: { type: 'circle' }, labels: { fill: colorScale[0] } },
-                    { name: toTitleCase(line2), symbol: { type: 'square' }, labels: { fill: colorScale[1] } },
+                    {
+                        name: toTitleCase(line1),
+                        symbol: { type: 'circle' },
+                        labels: { fill: graphStyles.colorScale[0] },
+                    },
+                    {
+                        name: toTitleCase(line2),
+                        symbol: { type: 'square' },
+                        labels: { fill: graphStyles.colorScale[1] },
+                    },
                 ]}
             />
         </VictoryChart>
@@ -193,56 +204,6 @@ const GraphLine = ({
 }
 
 export default GraphLine
-
-const colorScale = ['#e25a2d', '#787276']
-// e25a2d, alt orange #ed7014
-
-const baseLabelStyles = {
-    fontFamily: "'Helvetica Neue', 'Helvetica', sans-serif",
-    fontSize: 10,
-    letterSpacing: 'normal',
-    padding: 8,
-    fill: 'grey',
-    stroke: 'transparent',
-    strokeWidth: 0,
-}
-
-export const styles = {
-    title: {
-        fill: colorScale[0],
-        fontSize: 22,
-    },
-    axisZero: {
-        axis: { stroke: '#8d4004', strokeWidth: 1 },
-        ticks: { stroke: '#8d4004', strokeWidth: 1, size: 5 },
-        tickLabels: { ...baseLabelStyles, fill: '#8d4004' },
-    },
-    seriesOne: {
-        data: { stroke: colorScale[0], fillOpacity: 1, fill: colorScale[0], strokeWidth: 0 },
-        labels: baseLabelStyles,
-    },
-    lineOne: {
-        data: { stroke: colorScale[0], strokeWidth: 1, opacity: 0.7 },
-        labels: { ...baseLabelStyles, fill: colorScale[0] },
-    },
-    axisOne: {
-        axis: { stroke: colorScale[0], strokeWidth: 1 },
-        ticks: { stroke: colorScale[0], strokeWidth: 1, size: 5 },
-        tickLabels: { ...baseLabelStyles, fill: colorScale[0] },
-    },
-    seriesTwo: {
-        data: { stroke: colorScale[1], fillOpacity: 1, fill: colorScale[1], strokeWidth: 0 },
-        labels: { ...baseLabelStyles, fill: colorScale[1] },
-    },
-    lineTwo: {
-        data: { stroke: colorScale[1], strokeWidth: 1, opacity: 0.7 },
-        labels: { ...baseLabelStyles, fill: colorScale[1] },
-    },
-    axisTwo: {
-        ticks: { stroke: colorScale[1], strokeWidth: 1, size: 5 },
-        tickLabels: { ...baseLabelStyles, fill: colorScale[1] },
-    },
-}
 
 export const methods = {
     plotXYData: (journal, type, barChart = false) => {
