+++ src/components/graphsleep.js
@@ -12,6 +12,7 @@ import {
 } from 'victory'
 import { useAppContext } from '../services/context'
 import { sortByDate } from '../services/sort'
+import { graphStyles } from '../utils/styles'
 
 const GraphSleep = ({ title = undefined }) => {
     const { state } = useAppContext()
@@ -32,11 +33,7 @@ const GraphSleep = ({ title = undefined }) => {
 
                 Object.keys(user.journal).forEach(item => {
                     const journalDate = user.journal[item].journalDate
-                    // const jDate = journalDate.split('-')
-                    // const dt = new Date(jDate[0], jDate[1] - 1, jDate[2])
                     const dt = new Date(journalDate)
-                    // const graphDate =
-                    //     dt.toLocaleString('default', { month: 'short' }) + '-' + ('0' + jDate[2]).slice(-2)
                     const dow = weekday[dt.getDay()]
 
                     if (typeof dataset.overall[journalDate] === 'undefined') {
@@ -64,7 +61,6 @@ const GraphSleep = ({ title = undefined }) => {
             const gData = []
             const pData = []
             Object.keys(dataset.overall).forEach((key, idx) => {
-                // console.log(dataset.overall[key])
                 const avg = Number(dataset.overall[key]['total']) / Number(dataset.overall[key]['count'])
                 gData.push({ idx: idx, date: key, average: avg, dow: dataset.overall[key]['dow'] })
                 pData.push({ idx: idx, date: key, total: dataset.personal[key]['total'] })
@@ -89,7 +85,6 @@ const GraphSleep = ({ title = undefined }) => {
             .then(snapshot => {
                 const p = []
                 snapshot.forEach(item => {
-                    // console.log(item.val())
                     p.push(item.val())
                 })
                 calculateSleep(p)
@@ -109,8 +104,6 @@ const GraphSleep = ({ title = undefined }) => {
     Object.keys(myData).forEach((date, idx) => {
         maxY1 = myData[idx].total > maxY1 ? myData[idx].total : maxY1
     })
-    // console.log('graphsleep maxY1', maxY1)
-    // console.log('myData', myData)
 
     return (
         <VictoryChart
@@ -120,7 +113,9 @@ const GraphSleep = ({ title = undefined }) => {
             domainPadding={{ x: 0, y: 10 }}
             theme={VictoryTheme.material}
         >
-            {typeof title !== 'undefined' ? <VictoryLabel x={50} y={30} text={title} style={styles.title} /> : null}
+            {typeof title !== 'undefined' ? (
+                <VictoryLabel x={50} y={30} text={title} style={graphStyles.title} />
+            ) : null}
             <VictoryAxis
                 tickValues={xTickValues}
                 tickLabelComponent={<VictoryLabel text={({ datum }) => graphData[datum].dow || datum} />}
@@ -136,8 +131,8 @@ const GraphSleep = ({ title = undefined }) => {
                     y: [0, maxY1],
                 }}
             >
-                <VictoryLine interpolation={'natural'} style={styles.lineTwo} />
-                <VictoryScatter style={styles.seriesTwo} size={2} symbol="square" />
+                <VictoryLine interpolation={'natural'} style={graphStyles.lineTwo} />
+                <VictoryScatter style={graphStyles.seriesTwo} size={2} symbol="square" />
             </VictoryGroup>
 
             <VictoryGroup
@@ -151,7 +146,7 @@ const GraphSleep = ({ title = undefined }) => {
                 // y={d => d.total / maxY1}
             >
                 <VictoryLine
-                    style={styles.lineOne}
+                    style={graphStyles.lineOne}
                     interpolation={'natural'}
                     // labels={({ datum }) => datum.total.toFixed(1)}
                     labelComponent={
@@ -165,7 +160,7 @@ const GraphSleep = ({ title = undefined }) => {
                         />
                     }
                 />
-                <VictoryScatter style={styles.seriesOne} size={2} symbol="circle" />
+                <VictoryScatter style={graphStyles.seriesOne} size={2} symbol="circle" />
             </VictoryGroup>
 
             <VictoryLegend
@@ -175,15 +170,19 @@ const GraphSleep = ({ title = undefined }) => {
                 orientation="horizontal"
                 gutter={30}
                 style={{
-                    border: { stroke: colorScale[0], strokeWidth: 0 },
-                    title: { ...baseLabelStyles, fontSize: 14 },
-                    labels: { ...baseLabelStyles, fontSize: 14 },
+                    border: { stroke: graphStyles.colorScale[0], strokeWidth: 0 },
+                    title: { ...graphStyles.baseLabelStyles, fontSize: 14 },
+                    labels: { ...graphStyles.baseLabelStyles, fontSize: 14 },
                     data: { fillOpacity: 1 },
                 }}
-                colorScale={colorScale}
+                colorScale={graphStyles.colorScale}
                 data={[
-                    { name: 'My Sleep', labels: { fill: colorScale[0] } },
-                    { name: 'Overall Average', symbol: { type: 'square' }, labels: { fill: colorScale[1] } },
+                    { name: 'My Sleep', labels: { fill: graphStyles.colorScale[0] } },
+                    {
+                        name: 'Overall Average',
+                        symbol: { type: 'square' },
+                        labels: { fill: graphStyles.colorScale[1] },
+                    },
                 ]}
             />
         </VictoryChart>
@@ -191,53 +190,3 @@ const GraphSleep = ({ title = undefined }) => {
 }
 
 export default GraphSleep
-
-const colorScale = ['#e25a2d', '#787276']
-// e25a2d, alt orange #ed7014
-
-const baseLabelStyles = {
-    fontFamily: "'Helvetica Neue', 'Helvetica', sans-serif",
-    fontSize: 10,
-    letterSpacing: 'normal',
-    padding: 8,
-    fill: 'grey',
-    stroke: 'transparent',
-    strokeWidth: 0,
-}
-
-export const styles = {
-    title: {
-        fill: colorScale[0],
-        fontSize: 22,
-    },
-    axisZero: {
-        axis: { stroke: '#8d4004', strokeWidth: 1 },
-        ticks: { stroke: '#8d4004', strokeWidth: 1, size: 5 },
-        tickLabels: { ...baseLabelStyles, fill: '#8d4004' },
-    },
-    seriesOne: {
-        data: { stroke: colorScale[0], fillOpacity: 1, fill: colorScale[0], strokeWidth: 0 },
-        labels: { ...baseLabelStyles, fontSize: 10 },
-    },
-    lineOne: {
-        data: { stroke: colorScale[0], strokeWidth: 1, opacity: 0.7 },
-        labels: { ...baseLabelStyles, fill: colorScale[0], fontSize: 10 },
-    },
-    axisOne: {
-        axis: { stroke: colorScale[0], strokeWidth: 1 },
-        ticks: { stroke: colorScale[0], strokeWidth: 1, size: 5 },
-        tickLabels: { ...baseLabelStyles, fill: colorScale[0] },
-    },
-    seriesTwo: {
-        data: { stroke: colorScale[1], fillOpacity: 1, fill: colorScale[1], strokeWidth: 0 },
-        labels: { ...baseLabelStyles, fill: colorScale[1] },
-    },
-    lineTwo: {
-        data: { stroke: colorScale[1], strokeWidth: 1, opacity: 0.7 },
-        labels: { ...baseLabelStyles, fill: colorScale[1] },
-    },
-    axisTwo: {
-        ticks: { stroke: colorScale[1], strokeWidth: 1, size: 5 },
-        tickLabels: { ...baseLabelStyles, fill: colorScale[1] },
-    },
-}
